Rename csv download helper file
