+++ src/components/v2v3/V2V3Project/modals/V2V3DownloadActivityModal.tsx
@@ -11,7 +11,7 @@ import {
   downloadPayments,
   downloadRedemptions,
   downloadAdditionsToBalance,
-} from 'utils/buildActivityCsvs'
+} from 'utils/csvDownloadHelpers'
 import { DownloadOutlined } from '@ant-design/icons'
 
 export default function V2V3DownloadActivityModal(props: ModalProps) {
